change to double doesn't help

+++ triangles/cuda/include/aabb.h
@@ -14,11 +14,11 @@ public:
   __device__ point3 min() const { return minimum; }
   __device__ point3 max() const { return maximum; }
 
-  // __device__ bool hit(const ray &r, float t_min, float t_max) const {
+  // __device__ bool hit(const ray &r, double t_min, double t_max) const {
   //   for (int a = 0; a < 3; a++) {
-  //     float t0 = fmin((minimum[a] - r.origin()[a]) / r.direction()[a],
+  //     double t0 = fmin((minimum[a] - r.origin()[a]) / r.direction()[a],
   //                     (maximum[a] - r.origin()[a]) / r.direction()[a]);
-  //     float t1 = fmax((minimum[a] - r.origin()[a]) / r.direction()[a],
+  //     double t1 = fmax((minimum[a] - r.origin()[a]) / r.direction()[a],
   //                     (maximum[a] - r.origin()[a]) / r.direction()[a]);
   //     t_min = fmax(t0, t_min);
   //     t_max = fmin(t1, t_max);
@@ -30,14 +30,14 @@ public:
   //   return true;
   // }
 
-  __device__ inline bool hit(const ray &r, float t_min,
-                                   float t_max) const {
+  __device__ inline bool hit(const ray &r, double t_min,
+                                   double t_max) const {
     for (int a = 0; a < 3; a++) {
       auto invD = 1.0f / r.direction()[a];
       auto t0 = (min()[a] - r.origin()[a]) * invD;
       auto t1 = (max()[a] - r.origin()[a]) * invD;
-      if (invD < 0.0f){
-        float temp = t1;
+      if (invD < 0.0){
+        double temp = t1;
         t1 = t0;
         t0 = temp;
       }
