+++ triangles/cuda/include/texture.h
@@ -7,7 +7,7 @@
 
 class abstract_texture {
 public:
-  __device__ virtual color value(float u, float v, const point3 &p) const = 0;
+  __device__ virtual color value(double u, double v, const point3 &p) const = 0;
 };
 
 class solid_color : public abstract_texture {
@@ -15,10 +15,10 @@ public:
   __device__ solid_color() {}
   __device__ solid_color(color c) : color_value(c) {}
 
-  __device__ solid_color(float red, float green, float blue)
+  __device__ solid_color(double red, double green, double blue)
       : solid_color(color(red, green, blue)) {}
 
-  __device__ virtual color value(float u, float v,
+  __device__ virtual color value(double u, double v,
                                  const point3 &p) const override {
     return color_value;
   }
@@ -37,10 +37,10 @@ public:
   __device__ checker_texture(color c1, color c2)
       : even(new solid_color(c1)), odd(new solid_color(c2)) {}
 
-  __device__ virtual color value(float u, float v,
+  __device__ virtual color value(double u, double v,
                                  const point3 &p) const override {
-    float sines =
-        sinf(10.0f * p.x()) * sinf(10.0f * p.y()) * sinf(10.0f * p.z());
+    double sines =
+        sin(10.0 * p.x()) * sin(10.0f * p.y()) * sin(10.0 * p.z());
     if (sines < 0)
       return odd->value(u, v, p);
     else
@@ -56,22 +56,22 @@ class noise_texture : public abstract_texture {
 public:
   __device__ noise_texture() {}
 
-  __device__ noise_texture(float sc, curandState *local_rand_state)
+  __device__ noise_texture(double sc, curandState *local_rand_state)
       : scale(sc), noise(new perlin(local_rand_state)) {}
 
-  __device__ virtual color value(float u, float v,
+  __device__ virtual color value(double u, double v,
                                  const point3 &p) const override {
     // return color(1, 1, 1) * noise->noise(scale * p);
     // shift off integer values
     // return color(1, 1, 1) * 0.5 * (1.0 + noise->noise(scale * p));
     // return color(1, 1, 1) * noise->turb(scale * p);
-    return color(1.0f, 1.0f, 1.0f) * 0.5f *
-           (1.0f + sinf(scale * p.z() + 10.0f * noise->turb(scale * p)));
+    return color(1.0, 1.0, 1.0) * 0.5 *
+           (1.0 + sin(scale * p.z() + 10.0 * noise->turb(scale * p)));
   }
 
 public:
   perlin *noise;
-  float scale;
+  double scale;
 };
 
 class image_texture : public abstract_texture {
@@ -89,14 +89,14 @@ public:
 
   __device__ ~image_texture() { delete data; }
 
-  __device__ virtual color value(float u, float v,
+  __device__ virtual color value(double u, double v,
                                  const point3 &p) const override {
     if (data == nullptr) {
-      return color(0.0f, 1.0f, 1.0f);
+      return color(0.0, 1.0, 1.0);
     }
 
-    u = clamp(u, 0.0f, 1.0f);
-    v = 1.0f - clamp(v, 0.0f, 1.0f);
+    u = clamp(u, 0.0, 1.0);
+    v = 1.0f - clamp(v, 0.0, 1.0);
 
     int i = (int)(u * width);
     int j = (int)(v * height);
@@ -109,7 +109,7 @@ public:
     // try to shift the map
     i = (i + width / 2 + width / 3) % width;
 
-    const float color_scale = 1.0f / 255.0f;
+    const double color_scale = 1.0 / 255.0;
     auto pixel = data + j * bytes_per_scanline + i * bytes_per_pixel;
 
     return color(color_scale * pixel[0], color_scale * pixel[1],
