+++ triangles/cuda/include/vec3.h
@@ -9,17 +9,17 @@ using std::sqrt;
 class vec3 {
 public:
   __host__ __device__ vec3() : e{0, 0, 0} {}
-  __host__ __device__ vec3(float e0, float e1, float e2) : e{e0, e1, e2} {}
+  __host__ __device__ vec3(double e0, double e1, double e2) : e{e0, e1, e2} {}
 
-  __host__ __device__ inline float x() const { return e[0]; }
-  __host__ __device__ inline float y() const { return e[1]; }
-  __host__ __device__ inline float z() const { return e[2]; }
+  __host__ __device__ inline double x() const { return e[0]; }
+  __host__ __device__ inline double y() const { return e[1]; }
+  __host__ __device__ inline double z() const { return e[2]; }
 
   __host__ __device__ inline vec3 operator-() const {
     return vec3(-e[0], -e[1], -e[2]);
   }
-  __host__ __device__ inline float operator[](int i) const { return e[i]; }
-  __host__ __device__ inline float &operator[](int i) { return e[i]; }
+  __host__ __device__ inline double operator[](int i) const { return e[i]; }
+  __host__ __device__ inline double &operator[](int i) { return e[i]; }
 
   __host__ __device__ inline vec3 &operator+=(const vec3 &v) {
     e[0] += v.e[0];
@@ -42,27 +42,27 @@ public:
     return *this;
   }
 
-  __host__ __device__ inline vec3 &operator*=(const float t) {
+  __host__ __device__ inline vec3 &operator*=(const double t) {
     e[0] *= t;
     e[1] *= t;
     e[2] *= t;
     return *this;
   }
 
-  __host__ __device__ inline vec3 &operator/=(const float t) {
+  __host__ __device__ inline vec3 &operator/=(const double t) {
     return *this *= 1.0f / t;
   }
 
-  __host__ __device__ inline float length() const {
+  __host__ __device__ inline double length() const {
     return sqrt(length_squared());
   }
 
-  __host__ __device__ inline float length_squared() const {
+  __host__ __device__ inline double length_squared() const {
     return e[0] * e[0] + e[1] * e[1] + e[2] * e[2];
   }
 
 public:
-  float e[3];
+  double e[3];
 };
 
 // vec3 Utility Functions
@@ -83,19 +83,19 @@ __host__ __device__ inline vec3 operator*(const vec3 &u, const vec3 &v) {
   return vec3(u.e[0] * v.e[0], u.e[1] * v.e[1], u.e[2] * v.e[2]);
 }
 
-__host__ __device__ inline vec3 operator*(float t, const vec3 &v) {
+__host__ __device__ inline vec3 operator*(double t, const vec3 &v) {
   return vec3(t * v.e[0], t * v.e[1], t * v.e[2]);
 }
 
-__host__ __device__ inline vec3 operator*(const vec3 &v, float t) {
+__host__ __device__ inline vec3 operator*(const vec3 &v, double t) {
   return t * v;
 }
 
-__host__ __device__ inline vec3 operator/(vec3 v, float t) {
+__host__ __device__ inline vec3 operator/(vec3 v, double t) {
   return (1 / t) * v;
 }
 
-__host__ __device__ inline float dot(const vec3 &u, const vec3 &v) {
+__host__ __device__ inline double dot(const vec3 &u, const vec3 &v) {
   return u.e[0] * v.e[0] + u.e[1] * v.e[1] + u.e[2] * v.e[2];
 }
